% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countTable.R
\name{get_countTable}
\alias{get_countTable}
\title{Get count table for the specified column.}
\usage{
get_countTable(data, col)
}
\arguments{
\item{data}{A table}

\item{col}{A string}
}
\value{
Tibble with all value counts in target column.
}
\description{
Get count table for the specified column.
}
\examples{
data(iris)
get_countTable(iris,"Species")
}
