% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countTable.R
\docType{data}
\name{iris}
\alias{iris}
\title{Edgar Anderson's Iris Data}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Sepal.Length}{Sepal Length, in centimeter}
  \item{Sepal.Width}{Sepal Width of the diamond, in centimeter}
  \item{Petal.Length}{Petal Length, in centimeter}
  \item{Petal.Width}{Petal Width, in centimeter}
  \item{Species}{The species of the flower}
  ...
}}
\usage{
iris
}
\description{
This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica.
}
\keyword{datasets}
